+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v6.4-SNAPSHOT
 - Analyzer
   - Added WebviewAppNameVersion.
   - Added a 'not contains' operator '!~'
+  - Parser can handle a few more tricky cases
 - New/improved detections
   - More Apple devices
   - VivoBrowser often implies Vivo DeviceBrand